+++ libcc2rs/src/va_args.rs
